requirements.txt
# ── Core UI & Data ─────────────────────────────────────────────────────────
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0

# ── Visualization ───────────────────────────────────────────────────────────
plotly>=5.20.0
matplotlib>=3.8.0
seaborn>=0.13.0

# ── Machine Learning ────────────────────────────────────────────────────────
scikit-learn>=1.4.0
scipy>=1.12.0

# ── Database ────────────────────────────────────────────────────────────────
# sqlite3 is part of the Python standard library — no install needed

# ── Data Ingestion ──────────────────────────────────────────────────────────
kaggle>=1.6.12
requests>=2.31.0

# ── AI / LLM ────────────────────────────────────────────────────────────────
openai>=1.13.0
langchain>=0.1.12
langchain-openai>=0.0.8
pandasai>=2.0.0

# ── Data Profiling ──────────────────────────────────────────────────────────
ydata-profiling>=4.6.4

# ── Environment & Config ────────────────────────────────────────────────────
python-dotenv>=1.0.1

# ── Export ──────────────────────────────────────────────────────────────────
xlsxwriter>=3.2.0
openpyxl>=3.1.2
reportlab>=4.1.0

# ── Testing ─────────────────────────────────────────────────────────────────
pytest>=8.1.0
